requirements.txt
azure-ai-inference
azure-identity
azure-search-documents
opencv-python
Pillow
numpy
python-dotenv
